Restructured pages, renamed CustomLink component to Anchor, tweaked the Theme Switcher button on mobile screens
