+++ components/layouts/partials/Header.jsx
@@ -1,4 +1,4 @@
-import CustomLink from '~/components/CustomLink'
+import Anchor from '~/components/Anchor'
 import styles from '~/styles/header.module.css'
 import FullLogo from './FullLogo'
 import ThemeSwitcher from './ThemeSwitcher'
@@ -9,9 +9,9 @@ export default function Header() {
             <div className="container flex p-5">
                 <FullLogo />
                 <nav className="mr-auto ml-4 pl-4 border-l border-gray-400 flex items-center text-base">
-                    <CustomLink href="/pokedex">
-                        <a className="mr-5 px-2 py-1">Pokedex</a>
-                    </CustomLink>
+                    <Anchor href="/information">
+                        <a className="mr-5 px-2 py-1">Information</a>
+                    </Anchor>
                 </nav>
                 <div className="my-auto">
                 <ThemeSwitcher />
